Change "url" for each event to signup
I don't think the URL key is respected by
Google Calendar or used at all and doesn't
appear to be used by the Wix calendar but does
show up in iCal. It looks like a long link, but
it passes along the information so the signup
form is all populated.